implement basic global & mongodb error filters, enhance deleteUser functionality by checking existence of user before deleting the user

+++ src/module/user/user.service.ts
@@ -76,6 +76,12 @@ export class UserService {
   }
 
   async deleteUser(id: string): Promise<void> {
+    const user = await this.userModel.findById(id);
+
+    if (!user) {
+      throw new BadRequestException('Invalid user id given.');
+    }
+
     await this.userModel.findByIdAndDelete(id);
   }
 }
